srcCodeFichiersSemiStruct/alchemy/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
PyJWT>=2.0
